% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_influence}
\alias{diagnostics_influence}
\title{Influence Diagnostic}
\usage{
diagnostics_influence(score)
}
\arguments{
\item{score}{Score vector from orthogonal inference.}
}
\value{
A list with influence summaries.
}
\description{
Influence Diagnostic
}
