% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_base.R
\name{learner_base}
\alias{learner_base}
\title{Create a Base Learner Specification}
\usage{
learner_base(name, fit_fun = NULL, predict_fun = NULL, params = list())
}
\arguments{
\item{name}{Learner identifier.}

\item{fit_fun}{Optional training function.}

\item{predict_fun}{Optional prediction function.}

\item{params}{Optional named or unnamed list of learner parameters.}
}
\value{
An object of class \code{learner_base}.
}
\description{
Create a Base Learner Specification
}
