% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuis_spec.R
\name{nuis_spec}
\alias{nuis_spec}
\title{Create a Nuisance Specification}
\usage{
nuis_spec(outcome_model, treatment_model)
}
\arguments{
\item{outcome_model}{A learner specification for outcome nuisance estimation.}

\item{treatment_model}{A learner specification for treatment nuisance estimation.}
}
\value{
An object of class \code{nuis_spec}.
}
\description{
Create a Nuisance Specification
}
